Add in timer methods to a new timerProvider.
This doesn't add in any new functionality, but we have  a few timer
related methods floating around in `MetalsLanguageServer`, and I've
started to really notice how much stuff we have stuck in there. I'd like
to slowly start moving some stuff out, and these were easy to move out.
Plus some of other methods that use these I plan on moving out as well,
so it will be helpful to have these in their own class to make that
easier instead of passing them from `MetalMetalsLanguageServer` into
other classes.

I ended up deleting the `Timed` class as well since it wasn't being used
at all.
