--- metals/src/main/scala/scala/meta/internal/metals/Timed.scala
@@ -1,6 +1,0 @@
-package scala.meta.internal.metals
-
-case class Timed[T](value: T, timer: Timer)
-object Timed {
-  def apply[T](value: T, time: Time): Timed[T] = Timed(value, new Timer(time))
-}
